+++ contracts/interfaces/ISYM.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 interface ISYM {
-    function mint(address _to, uint256 _amount) external;
+    function mint(address _to, uint _amount) external;
 
-    function burn(uint256 _amount) external;
+    function burn(uint _amount) external;
 }
