+++ contracts/interfaces/ISYMRate.sol
@@ -5,13 +5,13 @@ interface ISYMRate {
     /*=== struct ===*/
 
     struct Rate {
-        uint256 startTime;
-        uint256 rate;
+        uint startTime;
+        uint rate;
     }
 
     /*=== function ===*/
 
-    function getSum(uint256 start, uint256 end) external view returns (uint256 sum);
+    function getSum(uint start, uint end) external view returns (uint sum);
 
-    function rates(uint256) external view returns (uint256 startTime, uint256 rate);
+    function rates(uint) external view returns (uint startTime, uint rate);
 }
