+++ contracts/interfaces/IVolumeTracker.sol
@@ -5,25 +5,25 @@ interface IVolumeTracker {
     /*=== struct ===*/
 
     struct Tier {
-        uint256 requirement;
-        uint256 rebateRatio;
+        uint requirement;
+        uint rebateRatio;
     }
 
     /*=== function ===*/
 
-    function claimWeeklyTradingFeeCoupon(uint256 _t) external;
+    function claimWeeklyTradingFeeCoupon(uint _t) external;
 
     function coupon() external view returns (address);
 
-    function logTrade(address _account, uint256 _volume) external;
+    function logTrade(address _account, uint _volume) external;
 
     function market() external view returns (address);
 
-    function rebateTiers(uint256) external view returns (uint256 requirement, uint256 rebateRatio);
+    function rebateTiers(uint) external view returns (uint requirement, uint rebateRatio);
 
     function settings() external view returns (address);
 
-    function userDailyVolume(address, uint256) external view returns (uint256);
+    function userDailyVolume(address, uint) external view returns (uint);
 
-    function userWeeklyVolume(address, uint256) external view returns (uint256);
+    function userWeeklyVolume(address, uint) external view returns (uint);
 }
